requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.0.19
langchain>=0.1.0
langchain_core>=0.1.0
langchain_openai>=0.0.5

# LLM providers
openai>=1.0.0

# Tool integrations
tavily-python>=0.2.3
bs4>=0.0.1
markdownify>=0.11.6
pydantic>=2.0.0

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0
langsmith>=0.0.75